fixed lab-2
